smart-move current buf to new tab if only 1 exist
When there's only 1 tab opened and we want to move_current_buf() there's
no point in asking the user which should be the target tab (the only
valid answer would be the current one, which wouldn't move anything).

Instead create a new tab autonomously and move the current buffer there.